Make the strict clang build clean on OpenBSD 7.9
Fix the conversions reported by -Wconversion/-Wsign-conversion/-Wsign-compare
that -Werror rejects:

* ops.c: use a size_t 'newoff' for the truncate/write offsets (after a
  non-negative check), size_t for read sizes, cast the size_t file offsets
  passed to pread/pwrite/mmap/ftruncate and the uid/gid sentinels, and mark
  the intentionally unused readdir parameters;
* lfile.c: int for snprintf lengths, cast the file size and the offsets passed
  to ftruncate/pread;
* util.c: cast pointer differences, st_blksize, st_size and the readlinkat
  result;
* ds.c: cast pointer differences and sysconf, with a real return check;
* desc.c: size_t for the checksum size and guard/cast the file size.

No behaviour change beyond rejecting negative sizes that could not occur for
the supported node types.

+++ desc.c
@@ -120,7 +120,8 @@ emuxfs_desc_chk_symlink_content(struct emuxfs_desc *desc, dind dev_index,
 	if ((lnksz = readlinkat(fd, path, lnkbuf, PATH_MAX - 1)) == -1)
 		return 1;
 
-	emuxfs_desc_chk_provided_content(desc, (uint8_t *)lnkbuf, lnksz, alg);
+	emuxfs_desc_chk_provided_content(desc, (uint8_t *)lnkbuf, (size_t)lnksz,
+	    alg);
 	return 0;
 }
 
@@ -163,7 +164,7 @@ emuxfs_desc_chk_meta(uint8_t *sum_out, const struct emuxfs_desc *desc,
     enum emuxfs_chk_alg_type alg_type)
 {
 	struct emuxfs_chk chk;
-	ssize_t chksz;
+	size_t chksz;
 	uint64_t u64h, u64le;
 
 	chksz = emuxfs_chk_size(alg_type);
